+++ src/main/java/controller/Controller.java
@@ -1,4 +1,4 @@
-package webserver;
+package controller;
 
 import domain.HttpRequest;
 import domain.HttpResponse;
