--- src/test/java/domain/HttpBodyTest.java
@@ -1,7 +1,0 @@
-package domain;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class HttpBodyTest {
-
-}
